@@ -1,8 +1,16 @@
 module github.com/mmcloughlin/avo
 
-go 1.11
+go 1.22.0
+
+toolchain go1.22.7
 
 require (
-	golang.org/x/arch v0.0.0-20190909030613-46d78d1859ac
-	golang.org/x/tools v0.0.0-20190914235951-31e00f45c22e
+	golang.org/x/arch v0.10.0
+	golang.org/x/sys v0.25.0
+	golang.org/x/tools v0.25.0
+)
+
+require (
+	golang.org/x/mod v0.21.0 // indirect
+	golang.org/x/sync v0.8.0 // indirect
 )
