@@ -1,4 +1,4 @@
-// +build ignore
+//go:build ignore
 
 package main
 
@@ -12,7 +12,8 @@ func main() {
 	TEXT("Labels", NOSPLIT, "func() uint64")
 	XORQ(RAX, RAX)
 	INCQ(RAX)
-	Label("neverused")
+	Label("never_used")
+	Label("consecutive_label_also_never_used")
 	INCQ(RAX)
 	INCQ(RAX)
 	INCQ(RAX)
