@@ -3,9 +3,9 @@ package gen
 import (
 	"errors"
 	"fmt"
-	"sort"
 	"strings"
 
+	"github.com/mmcloughlin/avo/internal/api"
 	"github.com/mmcloughlin/avo/internal/inst"
 	"github.com/mmcloughlin/avo/internal/prnt"
 	"github.com/mmcloughlin/avo/printer"
@@ -17,7 +17,7 @@ type mov struct {
 }
 
 // NewMOV generates a function that will auto-select the correct MOV instruction
-// based on operand types and and sizes.
+// based on operand types and sizes.
 func NewMOV(cfg printer.Config) Interface {
 	return GoFmt(&mov{cfg: cfg})
 }
@@ -29,7 +29,7 @@ func (m *mov) Generate(is []inst.Instruction) ([]byte, error) {
 	m.Printf("import (\n")
 	m.Printf("\t\"go/types\"\n")
 	m.NL()
-	m.Printf("\t\"%s/operand\"\n", pkg)
+	m.Printf("\t%q\n", api.ImportPath(api.OperandPackage))
 	m.Printf(")\n\n")
 
 	m.Printf("func (c *Context) mov(a, b operand.Op, an, bn int, t *types.Basic) {\n")
@@ -47,87 +47,105 @@ func (m *mov) Generate(is []inst.Instruction) ([]byte, error) {
 }
 
 func (m *mov) instruction(i inst.Instruction) {
-	f := flags(i)
-	sizes, err := formsizes(i)
+	tcs := typecases(i)
+	mfs, err := movforms(i)
 	if err != nil {
 		m.AddError(err)
 		return
 	}
-	for _, size := range sizes {
+	for _, mf := range mfs {
 		conds := []string{
-			fmt.Sprintf("an == %d", size.A),
-			fmt.Sprintf("bn == %d", size.B),
+			fmt.Sprintf("an == %d", opsize[mf.A]),
+			fmt.Sprintf("%s(a)", api.CheckerName(mf.A)),
+			fmt.Sprintf("bn == %d", opsize[mf.B]),
+			fmt.Sprintf("%s(b)", api.CheckerName(mf.B)),
 		}
-		for c, on := range f {
-			cmp := map[bool]string{true: "!=", false: "=="}
-			cond := fmt.Sprintf("(t.Info() & %s) %s 0", c, cmp[on])
-			conds = append(conds, cond)
+
+		for _, tc := range tcs {
+			typecase := fmt.Sprintf("(t.Info() & %s) %s %s", tc.Mask, tc.Op, tc.Value)
+			m.Printf("case %s && %s:\n", strings.Join(conds, " && "), typecase)
+			m.Printf("c.%s(a, b)\n", i.Opcode)
 		}
-		sort.Strings(conds)
-		m.Printf("case %s:\n", strings.Join(conds, " && "))
-		m.Printf("c.%s(a, b)\n", i.Opcode)
 	}
 }
 
 // ismov decides whether the given instruction is a plain move instruction.
 func ismov(i inst.Instruction) bool {
-	if i.AliasOf != "" || !strings.HasPrefix(i.Opcode, "MOV") {
+	// Ignore aliases.
+	if i.AliasOf != "" {
 		return false
 	}
+
+	// Accept specific move instruction prefixes.
+	prefixes := []string{"MOV", "KMOV", "VMOV"}
+	accept := false
+	for _, prefix := range prefixes {
+		accept = strings.HasPrefix(i.Opcode, prefix) || accept
+	}
+	if !accept {
+		return false
+	}
+
+	// Exclude some cases based on instruction descriptions.
 	exclude := []string{"Packed", "Duplicate", "Aligned", "Hint", "Swapping"}
 	for _, substring := range exclude {
 		if strings.Contains(i.Summary, substring) {
 			return false
 		}
 	}
+
 	return true
 }
 
-func flags(i inst.Instruction) map[string]bool {
-	f := map[string]bool{}
-	switch {
-	case strings.Contains(i.Summary, "Floating-Point"):
-		f["types.IsFloat"] = true
-	case strings.Contains(i.Summary, "Zero-Extend"):
-		f["types.IsInteger"] = true
-		f["types.IsUnsigned"] = true
-	case strings.Contains(i.Summary, "Sign-Extension"):
-		f["types.IsInteger"] = true
-		f["types.IsUnsigned"] = false
-	default:
-		f["types.IsInteger"] = true
-	}
-	return f
+type typecase struct {
+	Mask  string
+	Op    string
+	Value string
 }
 
-type movsize struct{ A, B int8 }
+func typecases(i inst.Instruction) []typecase {
+	switch {
+	case strings.Contains(i.Summary, "Floating-Point"):
+		return []typecase{
+			{"types.IsFloat", "!=", "0"},
+		}
+	case strings.Contains(i.Summary, "Zero-Extend"):
+		return []typecase{
+			{"(types.IsInteger|types.IsUnsigned)", "==", "(types.IsInteger|types.IsUnsigned)"},
+			{"types.IsBoolean", "!=", "0"},
+		}
+	case strings.Contains(i.Summary, "Sign-Extension"):
+		return []typecase{
+			{"(types.IsInteger|types.IsUnsigned)", "==", "types.IsInteger"},
+		}
+	default:
+		return []typecase{
+			{"(types.IsInteger|types.IsBoolean)", "!=", "0"},
+		}
+	}
+}
 
-func (s movsize) sortkey() uint16 { return (uint16(s.A) << 8) | uint16(s.B) }
+type movform struct{ A, B string }
 
-func formsizes(i inst.Instruction) ([]movsize, error) {
-	set := map[movsize]bool{}
+func movforms(i inst.Instruction) ([]movform, error) {
+	var mfs []movform
 	for _, f := range i.Forms {
 		if f.Arity() != 2 {
 			continue
 		}
-		s := movsize{
-			A: opsize[f.Operands[0].Type],
-			B: opsize[f.Operands[1].Type],
+		mf := movform{
+			A: f.Operands[0].Type,
+			B: f.Operands[1].Type,
 		}
-		if s.A < 0 || s.B < 0 {
+		if opsize[mf.A] < 0 || opsize[mf.B] < 0 {
 			continue
 		}
-		if s.A == 0 || s.B == 0 {
+		if opsize[mf.A] == 0 || opsize[mf.B] == 0 {
 			return nil, errors.New("unknown operand type")
 		}
-		set[s] = true
+		mfs = append(mfs, mf)
 	}
-	var ss []movsize
-	for s := range set {
-		ss = append(ss, s)
-	}
-	sort.Slice(ss, func(i, j int) bool { return ss[i].sortkey() < ss[j].sortkey() })
-	return ss, nil
+	return mfs, nil
 }
 
 var opsize = map[string]int8{
@@ -140,9 +158,14 @@ var opsize = map[string]int8{
 	"r32":   4,
 	"r64":   8,
 	"xmm":   16,
+	"ymm":   32,
+	"zmm":   64,
 	"m8":    1,
 	"m16":   2,
 	"m32":   4,
 	"m64":   8,
 	"m128":  16,
+	"m256":  32,
+	"m512":  64,
+	"k":     8,
 }
