@@ -1,7 +1,5 @@
 package load
 
-import "github.com/mmcloughlin/avo/internal/inst"
-
 // alias defines an opcode alias.
 type alias struct {
 	From string
@@ -95,36 +93,50 @@ func init() {
 	aliases = append(aliases, annoyingaliases...)
 }
 
-// extras is simply a list of extra instructions to add to the database.
-var extras = []*inst.Instruction{
-	// MOVLQZX does not appear in either x86 CSV or Opcodes, but does appear in stdlib assembly.
+// maskrequired is a set of AVX-512 opcodes where the mask register is required.
+// Usually the mask register can be omitted, in which case K0 is implied.
+var maskrequired = map[string]bool{
+	// Reference: https://github.com/golang/go/blob/4fd94558820100129b98f284e21b19fc27a99926/src/cmd/internal/obj/x86/asm6.go#L4219-L4240
 	//
-	// Reference: https://github.com/golang/go/blob/048c9164a0c5572df18325e377473e7893dbfb07/src/runtime/asm_amd64.s#L451-L453
+	//		// Checks to warn about instruction/arguments combinations that
+	//		// will unconditionally trigger illegal instruction trap (#UD).
+	//		switch p.As {
+	//		case AVGATHERDPD,
+	//			AVGATHERQPD,
+	//			AVGATHERDPS,
+	//			AVGATHERQPS,
+	//			AVPGATHERDD,
+	//			AVPGATHERQD,
+	//			AVPGATHERDQ,
+	//			AVPGATHERQQ:
+	//			// AVX512 gather requires explicit K mask.
+	//			if p.GetFrom3().Reg >= REG_K0 && p.GetFrom3().Reg <= REG_K7 {
+	//				if !avx512gatherValid(ctxt, p) {
+	//					return
+	//				}
+	//			} else {
+	//				if !avx2gatherValid(ctxt, p) {
+	//					return
+	//				}
+	//			}
+	//		}
 	//
-	//	TEXT ·reflectcall(SB), NOSPLIT, $0-32
-	//		MOVLQZX argsize+24(FP), CX
-	//		DISPATCH(runtime·call32, 32)
-	//
-	// Reference: https://github.com/golang/go/blob/048c9164a0c5572df18325e377473e7893dbfb07/src/cmd/internal/obj/x86/asm6.go#L1217
-	//
-	//		{AMOVLQZX, yml_rl, Px, opBytes{0x8b}},
-	//
-	// Reference: https://github.com/golang/go/blob/048c9164a0c5572df18325e377473e7893dbfb07/src/cmd/internal/obj/x86/asm6.go#L515-L517
-	//
-	//	var yml_rl = []ytab{
-	//		{Zm_r, 1, argList{Yml, Yrl}},
-	//	}
-	//
-	{
-		Opcode:  "MOVLQZX",
-		Summary: "Move with Zero-Extend",
-		Forms: []inst.Form{
-			{
-				Operands: []inst.Operand{
-					{Type: "m32", Action: inst.R},
-					{Type: "r64", Action: inst.W},
-				},
-			},
-		},
-	},
+	"VGATHERDPD": true,
+	"VGATHERQPD": true,
+	"VGATHERDPS": true,
+	"VGATHERQPS": true,
+	"VPGATHERDD": true,
+	"VPGATHERQD": true,
+	"VPGATHERDQ": true,
+	"VPGATHERQQ": true,
+
+	// Restriction applies to SCATTER instructions too.
+	"VPSCATTERDD": true,
+	"VPSCATTERDQ": true,
+	"VPSCATTERQD": true,
+	"VPSCATTERQQ": true,
+	"VSCATTERDPD": true,
+	"VSCATTERDPS": true,
+	"VSCATTERQPD": true,
+	"VSCATTERQPS": true,
 }
