@@ -5,6 +5,7 @@ import (
 
 	"github.com/mmcloughlin/avo/attr"
 	"github.com/mmcloughlin/avo/build"
+	"github.com/mmcloughlin/avo/buildtags"
 	"github.com/mmcloughlin/avo/printer"
 	"github.com/mmcloughlin/avo/reg"
 )
@@ -73,11 +74,64 @@ func TestConstraints(t *testing.T) {
 	ctx.ConstraintExpr("linux,386 darwin,!cgo")
 	ctx.ConstraintExpr("!noasm")
 
+	expect := []string{
+		"// Code generated by avo. DO NOT EDIT.",
+		"",
+	}
+	if buildtags.GoBuildSyntaxSupported() {
+		expect = append(expect,
+			"//go:build ((linux && 386) || (darwin && !cgo)) && !noasm",
+		)
+	}
+	if buildtags.PlusBuildSyntaxSupported() {
+		expect = append(expect,
+			"// +build linux,386 darwin,!cgo",
+			"// +build !noasm",
+		)
+	}
+	expect = append(expect, "")
+
+	AssertPrintsLines(t, ctx, printer.NewGoAsm, expect)
+}
+
+func TestAlignmentNoOperands(t *testing.T) {
+	ctx := build.NewContext()
+	ctx.Function("alignment")
+	ctx.SignatureExpr("func()")
+	ctx.ADDQ(reg.RAX, reg.RBX)
+	ctx.VMOVDQU(reg.Y4, reg.Y11)
+	ctx.VZEROUPPER()
+	ctx.ADDQ(reg.R9, reg.R13)
+	ctx.RET()
+
 	AssertPrintsLines(t, ctx, printer.NewGoAsm, []string{
 		"// Code generated by avo. DO NOT EDIT.",
 		"",
-		"// +build linux,386 darwin,!cgo",
-		"// +build !noasm",
+		"// func alignment()",
+		"TEXT ·alignment(SB), $0",
+		"\tADDQ    AX, BX",
+		"\tVMOVDQU Y4, Y11",
+		"\tVZEROUPPER",      // instruction with no alignment doesn't affect width
+		"\tADDQ    R9, R13", // retains alignment from above
+		"\tRET",
+		"",
+	})
+}
+
+func TestOpcodeSuffixes(t *testing.T) {
+	ctx := build.NewContext()
+	ctx.Function("suffixes")
+	ctx.SignatureExpr("func()")
+	ctx.VADDPD_RD_SAE_Z(reg.Z1, reg.Z2, reg.K1, reg.Z3)
+	ctx.ADDQ(reg.RAX, reg.RBX)
+
+	AssertPrintsLines(t, ctx, printer.NewGoAsm, []string{
+		"// Code generated by avo. DO NOT EDIT.",
+		"",
+		"// func suffixes()",
+		"TEXT ·suffixes(SB), $0",
+		"\tVADDPD.RD_SAE.Z Z1, Z2, K1, Z3",
+		"\tADDQ            AX, BX", // suffixes count towards alignment width
 		"",
 	})
 }
